torcontrol: Move TorControlReply, TorControlConnection and TorController to improve testability
